Improved functionality PatientController
fx

+++ HospitalSol/Hospital.BL/Patient/PatientService.cs
@@ -1,33 +1,33 @@
 using FluentValidation;
 using Hospital.Abstraction.Interfaces;
 using Hospital.BL.Patient.Validators;
+using Hospital.Common.Exceptions;
 using Hospital.Common.Models;
+using Hospital.Common.Models.Collection;
 using Hospital.Common.Models.Patient;
 
 //using Hospital.Common.Exceptions;
 
 namespace Hospital.BL.Patient;
 
-public class PatientService: IPatientService
+public class PatientService : IPatientService
 {
     private readonly IValidator<CreatePatientModel> _createPatientModelValidator;
-    //private readonly IValidator<UpdatePatientModel> _updatePatientModelValidator;
+    private readonly IValidator<UpdatePatientModel> _updatePatientModelValidator;
     private readonly IPatientRepository _patientRepository;
-    //private readonly IValidator<GetListModel<PatientFilterModel>> _getFilterModelValidator;
-    //private readonly IAreaService _areaService;
+    private readonly IValidator<GetListModel<PatientFilterModel>> _getFilterModelValidator;
+
 
     public PatientService(
         IValidator<CreatePatientModel> createPatientModelValidator,
-     //   IValidator<UpdatePatientModel> updatePatientModelValidator,
-       // IValidator<GetListModel<PatientFilterModel>> getFilterModelValidator,
-        IPatientRepository patientRepository)//,
-        //IAreaService areaService)
+        IValidator<UpdatePatientModel> updatePatientModelValidator,
+        IValidator<GetListModel<PatientFilterModel>> getFilterModelValidator,
+        IPatientRepository patientRepository)
     {
         _createPatientModelValidator = createPatientModelValidator;
-       // _updatePatientModelValidator = updatePatientModelValidator;
-        //_getFilterModelValidator = getFilterModelValidator;
+        _updatePatientModelValidator = updatePatientModelValidator;
+        _getFilterModelValidator = getFilterModelValidator;
         _patientRepository = patientRepository;
-        //_areaService = areaService;
     }
 
     /// <inheritdoc cref="IPatientService.CreateAsync(CreatePatientModel)"/>
@@ -35,7 +35,6 @@ public class PatientService: IPatientService
     {
         await _createPatientModelValidator.ValidateAndThrowAsync(createPatientModel);
         return await _patientRepository.CreateAsync(createPatientModel);
-
     }
     
     /// <inheritdoc cref="IPatientService.GetAsync(long)"/>
@@ -47,10 +46,20 @@ public class PatientService: IPatientService
     public async Task DeleteAsync(long patientId)
     {
         var patientModel = await _patientRepository.GetAsync(patientId);
-
-        //if (patientModel == null)
-         //   throw new EntityNotFoundException($"События с идентификатором {patientId} не существует");
-
+        if (patientModel == null)
+           throw new EntityNotFoundException($"Пациента с идентификатором {patientId} не существует");
         await _patientRepository.DeleteAsync(patientId);
+    }
+
+    public async Task UpdateAsync(UpdatePatientModel updatePatientModel)
+    {
+        await _updatePatientModelValidator.ValidateAndThrowAsync(updatePatientModel);
+        await _patientRepository.UpdateAsync(updatePatientModel);
+    }
+
+    public async Task<BaseCollectionModel<ListPatientModel>> GetAsync(GetListModel<PatientFilterModel> getListModel)
+    {
+        await _getFilterModelValidator.ValidateAndThrowAsync(getListModel);
+        return await _patientRepository.GetAsync(getListModel);
     }
 }
